Refactor: 템플릿 페이지 폴더 구조화 및 기본 경로 리다이렉트
- 기본 경로(/)를 /wrong-notes로 리다이렉트
- 템플릿 관련 페이지를 /templates 폴더로 이동
  - /template-tester → /templates
  - /template-registration → /templates/register
  - /my-templates → /templates/my
- 관련 navigate 경로 및 테스트 일괄 치환
