+++ Template-Tester_FE/src/pages/templates/register.tsx
